+++ eslint.config.cjs
@@ -4,6 +4,7 @@ const promise = require('eslint-plugin-promise');
 
 const rendererGlobals = Object.fromEntries([
   'MAX_PORTRAIT_VARIANTS', 'state', 'resetSelection', 'hasMaximumPortraits', 'modalPreserveAttribute',
+  'DEFAULT_CHARACTER_NAME_COLOR', 'DEFAULT_TITLE_GLOW_COLOR', 'CHARACTER_COLOR_PRESETS', 'normalizeAppearanceColor', 'hasReadableNameColor', 'applyCharacterAppearance', 'syncCharacterAppearance', 'showCharacterAppearanceModal', 'trapAppearanceFocus', 'updateAppearanceSelection', 'updateAppearancePreview', 'resetAppearanceEditor', 'saveCharacterAppearance',
   'morphAppContent', 'morphAppRegion',
   'chromeMarkup', 'contextMenuMarkup', 'cardMarkup', 'tableMarkup', 'sidebarMarkup', 'inspectorMarkup', 'mainMarkup', 'sortControlMarkup', 'filterPanelMarkup', 'emptyResultsMarkup',
   'showCropModal', 'initializeCropInteraction', 'resetCropPosition', 'clampCropOffset', 'applyCropTransform', 'saveCroppedPortrait', 'appendPortrait',